Separate test cleanup from tests runnning
https://eaflood.atlassian.net/browse/PSD-420

SonarQube needs the test results to record code coverage. Currently the
test results XML file is deleted after publishing to Jenkins.

This change is to prevent tests being deleted in the runTests function.
Test output will be deleted in its own function that can be run in a final cleanup step.