Remove redundant variable definitions
Because of the way groovy works under the hood the variable definitions
removed did not create globals and were redundant (see
[here](https://groovy-lang.org/structure.html#_methods for more detail)
)

This means that for, say, containerTag the assignment in getVariables is
what actually creates the global var.

I would argue that we shouldn't be using global vars and given the
method returns values this is perhaps not the intent.