Fix breaking change to helm repo add (#178)
Helm 3.3.2 has a breaking change:

The default behavior of helm repo add was changed: It no longer overwrites repositories by default. The flag --force-update was added to helm repo add to allow a repo to be overwritten. The flag --no-update still exists, but does nothing. This change accords with the Helm policy for allowing minor breaking changes in the name of security.

full details https://github.com/helm/helm/releases

Co-authored-by: Mark Harrop <mark.harrop@defra.gov.uk>
Co-authored-by: John Watson <johnwatson484@hotmail.com>